add authntuction app jwt

+++ graphql-prisma/src/resolvers/Mutation.js
@@ -1,9 +1,20 @@
-import uuidv4 from 'uuid/v4';
+import bcrypt from 'bcryptjs';
+import jwt from 'jsonwebtoken';
+import Auth from '../utils/Auth';
 
 const Mutation = {
+  async login(parent, args, { prisma }, info) {
+    const { email, password } = args.data;
+    if (!email || !password) throw new Error('There are required field missing');
+    const user = await prisma.query.user({ where: { email } });
+    if (!user) throw new Error('there are some error in your email or password');
+    const configPassword = await bcrypt.compare(password, user.password);
+    if (!configPassword) throw new Error('Make sure from your email and password');
+    return { user, token: jwt.sign({ userId: user.id }, 'thisissecretcode') };
+  },
   async createUser(parent, args, { prisma }, info) {
     const {
-      data: { email }
+      data: { email, password: beforeHashing }
     } = args;
     const { data: newUserData } = args;
 
@@ -12,48 +23,59 @@ const Mutation = {
       throw new Error('Email already taken');
     }
 
-    const user = await prisma.mutation.createUser({ data: newUserData }, info);
+    if (beforeHashing.length < 8) throw new Error('password must be 8 character or more.');
+
+    const password = await bcrypt.hash(beforeHashing, 12);
+
+    const user = await prisma.mutation.createUser({ data: { ...newUserData, password } });
     if (!user) {
       throw new Error('there are some problem in create user');
     }
-    return user;
+    return { user, token: jwt.sign({ userId: user.id }, 'thisissecretcode') };
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    const { id } = args;
-    const userExist = await prisma.exists.User({ id });
+
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userIdFromAuth = Auth(request);
+    const userExist = await prisma.exists.User({ id: userIdFromAuth });
     if (!userExist) throw new Error('this user not found');
 
-    const user = await prisma.mutation.deleteUser({ where: { id } }, info);
+    const user = await prisma.mutation.deleteUser({ where: { id: userIdFromAuth } }, info);
     return user;
   },
-  async updateUser(parent, args, { prisma }, info) {
-    const { id, data } = args;
-    const userExist = await prisma.exists.User({ id });
+
+  async updateUser(parent, args, { prisma, request }, info) {
+    const userIdFromAuth = Auth(request);
+    const { data } = args;
+    const userExist = await prisma.exists.User({ id: userIdFromAuth });
     if (!userExist) throw new Error('User you lookup not found');
     const dataKeys = Object.keys(data);
     if (dataKeys.includes('id')) throw new Error('You cant change that property');
     if (dataKeys.includes('posts' || 'comments'))
       throw new Error('You can not change from that place');
-    const updateUser = await prisma.mutation.updateUser({ data: data, where: { id } }, info);
+    const updateUser = await prisma.mutation.updateUser(
+      { data: data, where: { id: userIdFromAuth } },
+      info
+    );
     return updateUser;
   },
-  async createPost(parent, args, { prisma, pubsub }, info) {
+
+  async createPost(parent, args, { prisma, request }, info) {
+    const userIdFromAuth = Auth(request);
     const { data } = args;
-    const { title, body, published, author } = data;
+    const { title, body, published } = data;
     if (!title || !body || !published) {
       throw new Error('There are require field missing');
     }
-    if (!author) {
+    if (!userIdFromAuth) {
       throw new Error('No user to create post');
     }
-    const userExist = await prisma.exists.User({ id: author });
+    const userExist = await prisma.exists.User({ id: userIdFromAuth });
     if (!userExist) throw new Error('No User with this id');
-    delete data.author;
     const dataCreatePost = {
       ...data,
       author: {
         connect: {
-          id: author
+          id: userIdFromAuth
         }
       }
     };
@@ -61,20 +83,30 @@ const Mutation = {
     if (!post) throw new Error('there are problem in create post');
     return post;
   },
-  async deletePost(parent, args, { prisma, pubsub }, info) {
+
+  async deletePost(parent, args, { prisma, request }, info) {
     const { id } = args;
-    if (!id) throw new Error('there are require field is missing');
-    const post = await prisma.exists.Post({ id });
+    const userIdFromAuth = Auth(request);
+
+    const post = await prisma.query.post({ where: { id } }, `{id title author{id}}`);
     if (!post) throw new Error('no post !!');
+
+    if (userIdFromAuth !== post.author.id) throw new Error('You cant delete that error sorry');
+    if (!id) throw new Error('there are require field is missing');
 
     const postDeleted = await prisma.mutation.deletePost({ where: { id } }, info);
     return postDeleted;
   },
-  async updatePost(parent, args, { prisma, pubsub }, info) {
+
+  async updatePost(parent, args, { prisma, request }, info) {
     const { id, data } = args;
-    const post = await prisma.exists.Post({ id });
+    const userIdFromAuth = Auth(request);
+
+    const post = await prisma.query.post({ where: { id } }, `{id title author{id}}`);
+
     if (!post) throw new Error('no user found with this id sorry');
-    // const originalPost = { ...post };
+    if (post.author.id !== userIdFromAuth)
+      throw new Error('you have not permission to update that');
     const dataKeys = Object.keys(data);
     if (dataKeys.includes('author' || 'id')) throw new Error('You cant change that property');
     if (dataKeys.includes('comments')) throw new Error('You can change from the place ');
@@ -82,28 +114,28 @@ const Mutation = {
     const postUpdated = await prisma.mutation.updatePost({ data: data, where: { id } }, info);
     return postUpdated;
   },
-  async createComment(parent, args, { prisma, pubsub }, info) {
+
+  async createComment(parent, args, { prisma, request }, info) {
     const { data } = args;
-    const { text, author, post: postId } = data;
+    const { text, post: postId } = data;
+    const userIdFromAuth = Auth(request);
     if (!text) {
       throw new Error('there are required field missing');
-    }
-    if (!author) {
-      throw new Error('no user to add comment');
     }
     if (!postId) {
       throw new Error('no post to add comment');
     }
-    const userExist = await prisma.exists.User({ id: author });
+    const userExist = await prisma.exists.User({ id: userIdFromAuth });
     if (!userExist) throw new Error('This user is not exist');
     const postExist = await prisma.exists.Post({ id: postId });
+    console.log(postExist);
     if (!postExist) throw new Error('this post not valid');
 
     const dataCreateComment = {
       ...data,
       author: {
         connect: {
-          id: author
+          id: userIdFromAuth
         }
       },
       post: {
@@ -115,17 +147,29 @@ const Mutation = {
     const comment = await prisma.mutation.createComment({ data: dataCreateComment }, info);
     return comment;
   },
-  async deleteComment(parent, args, { prisma, pubsub }, info) {
+
+  async deleteComment(parent, args, { prisma, request }, info) {
     const { id } = args;
+    const userIdFromAuth = Auth(request);
+
     if (!id) throw new Error('no comment to delete it');
-    const comment = await prisma.exists.Comment({ id });
+
+    const comment = await prisma.query.comment({ where: { id } }, `{author{id}}`);
     if (!comment) throw new Error('Sorry no post to delete it');
+    if (comment.author.id !== userIdFromAuth)
+      throw new Error('you have not permission to deleted that comment');
+
     return prisma.mutation.deleteComment({ where: { id } }, info);
   },
-  async updateComment(parent, args, { prisma, pubsub }, info) {
+
+  async updateComment(parent, args, { prisma, request }, info) {
     const { id, data } = args;
-    const comment = await prisma.exists.Comment({ id });
+    const userIdFromAuth = Auth(request);
+
+    const comment = await prisma.query.comment({ where: { id } }, `author{id}`);
     if (!comment) throw new Error('no comment found with this id sorry');
+    if (comment.author.id !== userIdFromAuth)
+      throw new Error('You have not permission to update that comment');
     const dataKeys = Object.keys(data);
     if (dataKeys.includes('author' || 'id' || 'posts'))
       throw new Error('You cant change that property');
